+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 ## [Unreleased]
 
+## [0.7.0] - 2024-02-08
+
+- Improvement in steroid recipe: migration
+  - Create new migration interactively for modify table with
+    * add_column, remove_column
+    * add_index, remove_index
+    * add_remove, remove_reference
+
 ## [0.6.0] - 2024-02-07
 
-- New steroid recipe: migration (Create new migration interactively)
+- New steroid recipe: migration
+  - Create new migration interactively for
+    * Create table
+    * Create join table
+    * Drop table
 
 ## [0.5.0] - 2024-02-06